requirements.txt
asgiref==3.7.2
Django~=5.0.4
sqlparse==0.5.0
python-decouple==3.8
django-taggit==5.0.1
Markdown==3.6
psycopg==3.1.18
djangorestframework==3.15.2
django-filter==24.3